requirements.txt
openai>=1.0.0
python-dotenv>=0.19.0
aiohttp>=3.8.0
async-timeout>=4.0.0
typing-extensions>=4.0.0 